refactor(shared): turn into an encapsulated module
This way it will be quite easy to import from Vis Util later on.
