+++ lib/index-legacy.ts
@@ -29,7 +29,7 @@ export { data };
 export { DataSet, DataView, Queue } from "vis-data/esnext";
 
 // bundled external libraries
-import * as Hammer from "./module/hammer";
+import { Hammer } from "./shared";
 export { Hammer };
 import * as keycharm from "keycharm";
 export { keycharm };
