+++ lib/network/Network.js
@@ -9,7 +9,7 @@ import {
 } from "vis-util/esnext";
 import { DOTToGraph } from "./dotparser";
 import { parseGephi } from "./gephiParser";
-import { Activator } from "../shared/Activator";
+import { Activator } from "../shared";
 import * as locales from "./locales";
 import { normalizeLanguageCode } from "./locale-utils";
 
@@ -26,8 +26,8 @@ import InteractionHandler from "./modules/InteractionHandler";
 import SelectionHandler from "./modules/SelectionHandler";
 import LayoutEngine from "./modules/LayoutEngine";
 import ManipulationSystem from "./modules/ManipulationSystem";
-import { Configurator } from "../shared/Configurator";
-import { VALIDATOR_PRINT_STYLE, Validator } from "../shared/Validator";
+import { Configurator } from "../shared";
+import { VALIDATOR_PRINT_STYLE, Validator } from "../shared";
 import { allOptions, configureOptions } from "./options.js";
 import KamadaKawai from "./modules/KamadaKawai.js";
