+++ lib/network/modules/Canvas.js
@@ -1,4 +1,4 @@
-import Hammer from "../../module/hammer";
+import { Hammer } from "../../shared";
 import { onRelease, onTouch } from "../../hammerUtil";
 
 import {
