+++ lib/network/modules/InteractionHandler.js
@@ -6,7 +6,7 @@ import {
   selectiveNotDeepExtend,
 } from "vis-util/esnext";
 import NavigationHandler from "./components/NavigationHandler";
-import { Popup } from "../../shared/Popup";
+import { Popup } from "../../shared";
 
 /**
  * Handler for interactions
