+++ lib/network/modules/ManipulationSystem.js
@@ -2,7 +2,7 @@ import "./ManipulationSystem.css";
 
 import { deepExtend, recursiveDOMDelete } from "vis-util/esnext";
 import { v4 as randomUUID } from "uuid";
-import Hammer from "../../module/hammer";
+import { Hammer } from "../../shared";
 import { onTouch } from "../../hammerUtil";
 
 /**
