+++ lib/network/modules/components/NavigationHandler.js
@@ -1,6 +1,6 @@
 import "./NavigationHandler.css";
 
-import Hammer from "../../../module/hammer";
+import { Hammer } from "../../../shared";
 import { onRelease, onTouch } from "../../../hammerUtil";
 import keycharm from "keycharm";
