+++ lib/shared/activator.js
@@ -1,7 +1,7 @@
-import "./Activator.css";
+import "./activator.css";
 
 import Emitter from "component-emitter";
-import Hammer from "../module/hammer";
+import { Hammer } from "./hammer";
 
 /**
  * Turn an element into an clickToUse element.
