+++ lib/shared/color-picker.css
