+++ lib/shared/color-picker.js
@@ -1,6 +1,6 @@
-import "./ColorPicker.css";
+import "./color-picker.css";
 
-import Hammer from "../module/hammer";
+import { Hammer } from "./hammer";
 import {
   HSVToRGB,
   RGBToHSV,
