+++ lib/shared/configurator.css
