+++ lib/shared/configurator.js
@@ -1,8 +1,8 @@
-import "./Configurator.css";
+import "./configurator.css";
 
 import { copyAndExtendArray } from "vis-util/esnext";
 
-import { ColorPicker } from "./ColorPicker";
+import { ColorPicker } from "./color-picker";
 
 /**
  * Wrap given text (last argument) in HTML elements (all preceding arguments).
