+++ lib/shared/hammer.js
@@ -32,4 +32,4 @@ const Hammer =
         return hammerMock();
       };
 
-export default Hammer;
+export { Hammer };
