+++ lib/shared/popup.js
@@ -1,4 +1,4 @@
-import "./Popup.css";
+import "./popup.css";
 
 /**
  * Popup is a class to create a popup window with some text
