+++ test/Configurator.test.js
@@ -1,7 +1,7 @@
 import assert from "assert";
 
 import { canvasMockify } from "./canvas-mock";
-import { Configurator } from "../lib/shared/Configurator";
+import { Configurator } from "../lib/shared";
 import Network from "../lib/network/Network";
 import { configureOptions } from "../lib/network/options.js";
